Added a small downloading-script and removed HoN-files
So HoN is no longer required to build this image.
To download HoN we now download a manifest-file, extract and parse it
and download a couple of files with urls generated from its content.
Those files are enough to run the repair-tool of HoN, which should
download the remaining files.